% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A \code{data.frame} which includes 3 columns, \code{v1} \code{v2}
and \code{w}. Where \code{v1} and \code{v2} are vector nodes with
their corresponding weight in column \code{w}}

\item{init_node}{The node where the search starts from}
}
\value{
The cost to each node from the start node
}
\description{
This is a description
}
\references{
https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm
}
