% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidian.R
\name{euclidean}
\alias{euclidean}
\title{Euclidian algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{numeric value}

\item{b}{numeric value}
}
\value{
A GCD value thats numeric
}
\description{
Function that computes GCD value between two numeric values
}
\examples{
euclidean(123612, 13892347912)
euclidean(101, 10612)
}
\references{
https://en.wikipedia.org/wiki/Euclidean
}
