% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki graph}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{Node where the edge starts}
  \item{v2}{Node where the edge ends}
  \item{w}{Weight of the edge}
}
}
\usage{
wiki_graph
}
\description{
Example graph from the Wikipedia page on Dijkstra's algorithm.
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\keyword{datasets}
