% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A \code{data.frame} which includes 3 columns, \code{v1} \code{v2}
and \code{w}. Where \code{v1} and \code{v2} are vector nodes with
their corresponding weight in column \code{w}}

\item{init_node}{The node where the search starts from}
}
\value{
The cost to each node from the start node
}
\description{
Algorithm used to find the shortest path from a starting node to every
other node in a weighted graph. It repeatedly picks the unvisited node with
the lowest cost and updates the cost of its neighbours until all nodes have
been visited.
}
\examples{
dijkstra(wiki_graph, 3)
dijkstra(wiki_graph, 1)
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
