% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Greatest Common Divisor (Euclidean Algorithm)}
\usage{
euclidean(a, b)
}
\arguments{
\item{a, b}{Numeric scalars}
}
\value{
Integer scalar: the Greatest Common Divisor of a and b.
}
\description{
Compute the greatest common divisor of two integers
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
