% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki Graph Dataset}
\format{
A data frame with 3 columns:
\describe{
\item{v1}{The starting node (numeric).}
\item{v2}{The ending node (numeric).}
\item{w}{The weight of the edge between the nodes (numeric).}
}
}
\source{
Data derived from the Dijkstra algorithm example on Wikipedia.
}
\usage{
wiki_graph
}
\description{
This dataset represents a graph used for demonstrating Dijkstra's algorithm.
It contains nodes (vertices) and edges (connections between nodes) with weights.
}
\keyword{datasets}
