% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Weighted edge list from Wikipedia example}
\format{
A data frame with three columns:
\describe{
  \item{v1}{From node (numeric)}
  \item{v2}{To node (numeric)}
  \item{w}{Edge weight (non-negative numeric)}
}
}
\usage{
data(wiki_graph)
}
\description{
A small weighted directed graph used in Dijkstra examples.
}
\examples{
data(wiki_graph)
head(wiki_graph)
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
\keyword{datasets}
